#2 测试html显示node数据14

+++ nodeRun.sh
@@ -3,13 +3,16 @@
 #description:nodeRun service
 #检测文件内容是否更改
 # 检测的文件
-package=/www/server/server.js
-# package=./views/index.html
+# while true
+# do
+# package=/www/server/server.js
+
+package=./server.js
 # 记录 md5值的文件
 md5=package_md5
 # 创建新的md5信息
-while true
-do
+
+# cd /www/server/
 package_md5_new=$(md5sum -b $package | awk '{print $1}'|sed 's/ //g')
 
 # 创建md5的函数
@@ -30,28 +33,42 @@ package_md5_old=$(cat $md5|sed 's/ //g')
 
 echo $package_md5_new
 echo $package_md5_old
+PIDS=`lsof -i:8088 | wc -l`
+echo $PIDS
+if [ "$PIDS" -gt "0" ]; then
+echo '启动了'
+else  
+echo '没启动'
+fi
 
 # 对象对比判断
 if [ "$package_md5_new" == "$package_md5_old" ];then
         echo 'md5 is not changed'
         echo '没有更改么，那我就运行了。。。'
-        cd /www/server
-        node $package
-        # docker restart saas
+        if [ "$PIDS" -gt "0" ]; then
+          echo '已经启动了'
+        else  
+          echo '开始启动了'
+          node $package
+        fi
+        docker restart saas
 else
         echo "md5 is  changed"
-        cd /www/server
-        echo "停止node服务"
-        kill -9 $(ps aux | grep server | awk '{print $2}')
-        echo "开始下载依赖"
-        # mkdir aa
-        npm i
-        echo "依赖下载完成重新开始服务"
-        node $package
+        if [ "$PIDS" -gt "0" ]; then
+          echo '已经启动了，让我关掉他'
+          kill -9 $(ps aux | grep server | awk '{print $2}')
+            echo "停止node服务"
+            echo "开始下载依赖"
+          # mkdir aa
+          npm i
+            echo "依赖下载完成重新开始服务"
+        fi
+        # cd /www/server/
+        echo '更改了，我要启动了'
         creatmd5
-        
+        node $package
 fi
-sleep 10;
+# sleep 10;
 
-done
+# done
